#1492 restructured packages
