+++ pkg/api/utils/apiServiceUtils.go
@@ -1,4 +1,4 @@
-package utils
+package api
 
 import (
 	"bytes"
