+++ pkg/api/utils/authUtils.go
@@ -1,4 +1,4 @@
-package utils
+package api
 
 import (
 	"net/http"
