+++ pkg/api/utils/eventUtils.go
@@ -1,4 +1,4 @@
-package utils
+package api
 
 import (
 	"encoding/json"
