+++ pkg/lib/events.go
@@ -9,6 +9,7 @@ import (
 	cloudeventshttp "github.com/cloudevents/sdk-go/pkg/cloudevents/transport/http"
 	"github.com/cloudevents/sdk-go/pkg/cloudevents/types"
 	"github.com/google/uuid"
+	"github.com/keptn/go-utils/pkg/api/models"
 	"log"
 	"net/url"
 	"time"
@@ -506,6 +507,10 @@ func (k *Keptn) SendTestsFinishedEvent(incomingEvent *cloudevents.Event, teststr
 	log.Println(fmt.Printf("%s", event))
 
 	return k.sendCloudEvent(event)
+}
+
+func (k *Keptn) GetEvents(eventType string) (*models.KeptnContextExtendedCE, *models.Error) {
+	return k.eventHandler.GetEvent(k.KeptnContext, eventType)
 }
 
 //
